added file helper and image validator

+++ FamilyTreeData/src/Repositories/Base/BaseRepository.cs
@@ -26,7 +26,7 @@ public class BaseRepository<T> : IBaseRepository<T> where T : class
         return await _dbSet.ToListAsync();
     }
 
-    public async Task<T?> GetByIdAsync(long id)
+    public async Task<T?> GetByIdAsync(int id)
     {
         return await _dbSet.FindAsync(id);
     }
